[ADD] add APIs in feign

See merge request hzero-c7ncd/devops-service!1787

+++ src/main/java/io/choerodon/devops/app/service/impl/UserAttrServiceImpl.java
@@ -179,7 +179,7 @@ public class UserAttrServiceImpl implements UserAttrService {
     }
 
     @Override
-    public long allUserCount() {
+    public int allUserCount() {
         return userAttrMapper.selectCount(new UserAttrDTO());
     }
